adapted code to process VB17 site data
B1+ maps acquired at the VB17 site (currently labelled as "NYU") do not have any mention of the reference voltage in their DICOM headers and .json files. Furthermore, the turboFLASH scans do not have an associated anatomical image. I have hardcoded the reference voltages used for all subjects at this site and modified the code so that the turboFLASH flip angle map is used for registration and not the anatomical that would normally be present for VE12 sites.